% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{iterations}

\item{n}{the sample size}

\item{p}{probability}
}
\value{
A bar plot and a table
}
\description{
Binomial Simulation
}
\examples{
mybin(iter = 100, n =10, p=0.5)
}
