% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Boostrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of Iterations (default: 10000)}

\item{x}{Sample data for bootstrapping}

\item{fun}{Function to apply to bootstrap samples (default: "mean")}

\item{alpha}{Significance level for confidence interval (default: 0.05)}

\item{cx}{Scaling factor for text size in the plot (default: 1.5)}

\item{...}{Additional graphical parameters for histogram}
}
\value{
A list with confidence interval (`ci`), function used (`fun`), and input data (`x`)
}
\description{
Computes bootstrap confidence intervals and visualizes the distribution of sample statistics.
}
\examples{
set.seed(39); sam=rnorm(25,mean=25,sd=10)
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
