% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{the 95 Percent Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{A numeric vector representing the sample.}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds of the confidence interval.
}
\description{
This function calculates a 95 Percent confidence interval for the mean of a single sample.
}
\examples{
myci(rnorm(30, mean=10, sd=12))
}
