% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{size}

\item{iter}{iteration}

\item{lambda}{The rate parameter of the Poisson distribution.}

\item{...}{}
}
\value{
A histogram of the sample means, a bar plot of the sampled values,
and a plot of the Poisson probability function.
}
\description{
Central Limit Theorem Poisson
}
\examples{
mycltp(n=10,iter=10000)
}
