% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Estimation for Beta Distribution}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{Numeric vector representing the sample data.}

\item{alpha}{Numeric vector of alpha values to evaluate.}

\item{beta}{Numeric vector of beta values to evaluate.}

\item{...}{Additional graphical parameters for contour plot.}
}
\value{
A list with sample vector (`x`), coordinates of max likelihood,
  max likelihood value (`maxl`), and estimated alpha (`maxalpha`)
  and beta (`maxbeta`).
}
\description{
Computes the maximum likelihood estimates of alpha and beta parameters
for a given sample vector.
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
