% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the distribution.}

\item{a}{The point up to which the area is calculated and shaded.}
}
\value{
A plot that shows a shaded area and the area itself
}
\description{
myncurve
}
\examples{
myncurve(5,10,3)
}
