% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{P-Value for a T-Test}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{Numeric, the observed t-statistic.}

\item{xmax}{Numeric, the maximum x-axis limit for the plot. Default is 4.}

\item{n}{Integer, the sample size. Default is 20.}

\item{alpha}{Numeric, the significance level for the t-test. Default is 0.05.}
}
\value{
A list containing:
  \item{q}{The critical t-value at the specified significance level.}
  \item{pvalue}{The calculated two-tailed p-value for the t-statistic.}
}
\description{
This function calculates the two-tailed p-value for a given t-statistic and plots the t-distribution with shaded areas
representing the critical regions. It can be used to visualize the relationship between the t-statistic, p-value,
and significance level.
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
tcalc=(mean(x1)-23)/(sd(x1)/sqrt(30))
mypvalue(tcalc,n=30,alpha=0.05)
}
