% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{The Square Function}
\usage{
mysquare(x)
}
\arguments{
\item{x}{# vector of quantitative data}
}
\value{
# vector
}
\description{
The Square Function
}
\examples{
mysquare(x = 1:10)
}
