% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{The ntickets Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Numbers of seats}

\item{gamma}{The Probability of overbooking}

\item{p}{The Probability of a "show"}
}
\value{
Two Plots of discrete and continuous graphs,
numbers of tickets for discrete and continuous values,
N value, p value, and gamma value
}
\description{
Calculate the number of tickets to be sold when the number of seats in flight is N and the probability of a
"show" is p and gamma is the probability that the plane will be overbooked.
}
\examples{
ntickets(200, 0.02, 0.95)
}
